Alteração no tipo de retorno da função obterMedicao

+++ imp/coletor_umidade_solo/recursos/SensorUmidadeSolo.h
@@ -10,7 +10,7 @@ class SensorUmidadeSolo
 {
 	public:
 		SensorUmidadeSolo(int lacos, int intervalo);
-		float obterMedicao(void);
+		int obterMedicao(void);
 
 	private:
 		int lacos;
